Add grpc protobuf

+++ receiver/message/message.go
@@ -2,6 +2,6 @@ package message
 
 type Message struct {
 	Text      string
-	AuthorId  int
-	Timestamp int
+	AuthorId  int64
+	Timestamp int32
 }
